:wrench: Moved `RigDef::Keyword` enum to common header.
Needed to break circular header dependencies, a nasty C++ thing. Factored out from a more complex buggy commit.
