+++ source/main/resources/rig_def_fileformat/RigDef_File.cpp
@@ -171,129 +171,6 @@ void Animation::AddMotorSource(unsigned int source, unsigned int motor)
     }
 }
 
-const char * KeywordToString(Keyword keyword)
-{
-    // PLEASE maintain alphabetical order!
-
-    switch (keyword)
-    {
-        case Keyword::ADD_ANIMATION:        return "add_animation";
-        case Keyword::AIRBRAKES:            return "airbrakes";
-        case Keyword::ANIMATORS:            return "animators";
-        case Keyword::ANTILOCKBRAKES:       return "antilockbrakes";
-        case Keyword::AUTHOR:               return "author";
-        case Keyword::AXLES:                return "axles";
-        case Keyword::BEAMS:                return "beams";
-        case Keyword::BRAKES:               return "brakes";
-        case Keyword::CAB:                  return "cab";
-        case Keyword::CAMERAS:              return "cameras";
-        case Keyword::CAMERARAIL:           return "camerarail";
-        case Keyword::CINECAM:              return "cinecam";
-        case Keyword::COLLISIONBOXES:       return "collisionboxes";
-        case Keyword::COMMANDS:             return "commands";
-        case Keyword::COMMANDS2:            return "commands2";
-        case Keyword::COMMENT:              return "comment";
-        case Keyword::CONTACTERS:           return "contacters";
-        case Keyword::CRUISECONTROL:        return "cruisecontrol";
-        case Keyword::DESCRIPTION:          return "description";
-        case Keyword::DETACHER_GROUP:       return "detacher_group";
-        case Keyword::DISABLEDEFAULTSOUNDS: return "disabledefaultsounds";
-        case Keyword::ENABLE_ADVANCED_DEFORMATION: return "enable_advanced_deformation";
-        case Keyword::END:                  return "end";
-        case Keyword::END_COMMENT:          return "end_comment";
-        case Keyword::END_DESCRIPTION:      return "end_description";
-        case Keyword::END_SECTION:          return "end_section";
-        case Keyword::ENGINE:               return "engine";
-        case Keyword::ENGOPTION:            return "engoption";
-        case Keyword::ENGTURBO:             return "engturbo";
-        case Keyword::EXHAUSTS:             return "exhausts";
-        case Keyword::EXTCAMERA:            return "extcamera";
-        case Keyword::FILEINFO:             return "fileinfo";
-        case Keyword::FILEFORMATVERSION:    return "fileformatversion";
-        case Keyword::FIXES:                return "fixes";
-        case Keyword::FLARES:               return "flares";
-        case Keyword::FLARES2:              return "flares2";
-        case Keyword::FLARES3:              return "flares3";
-        case Keyword::FLAREGROUPS_NO_IMPORT:return "flaregroups_no_import";
-        case Keyword::FLEXBODIES:           return "flexbodies";
-        case Keyword::FLEXBODY_CAMERA_MODE: return "flexbody_camera_mode";
-        case Keyword::FLEXBODYWHEELS:       return "flexbodywheels";
-        case Keyword::FORSET:               return "forset";
-        case Keyword::FORWARDCOMMANDS:      return "forwardcommands";
-        case Keyword::FUSEDRAG:             return "fusedrag";
-        case Keyword::GLOBALS:              return "globals";
-        case Keyword::GUID:                 return "guid";
-        case Keyword::GUISETTINGS:          return "guisettings";
-        case Keyword::HELP:                 return "help";
-        case Keyword::HIDEINCHOOSER:        return "hideinchooser";
-        case Keyword::HOOKGROUP:            return "hookgroup";
-        case Keyword::HOOKS:                return "hooks";
-        case Keyword::HYDROS:               return "hydros";
-        case Keyword::IMPORTCOMMANDS:       return "importcommands";
-        case Keyword::INTERAXLES:           return "interaxles";
-        case Keyword::LOCKGROUPS:           return "lockgroups";
-        case Keyword::LOCKGROUP_DEFAULT_NOLOCK: return "lockgroup_default_nolock";
-        case Keyword::MANAGEDMATERIALS:     return "managedmaterials";
-        case Keyword::MATERIALFLAREBINDINGS: return "materialflarebindings";
-        case Keyword::MESHWHEELS:           return "meshwheels";
-        case Keyword::MESHWHEELS2:          return "meshwheels2";
-        case Keyword::MINIMASS:             return "minimass";
-        case Keyword::NODES:                return "nodes";
-        case Keyword::NODES2:               return "nodes2";
-        case Keyword::PARTICLES:            return "particles";
-        case Keyword::PISTONPROPS:          return "pistonprops";
-        case Keyword::PROP_CAMERA_MODE:     return "prop_camera_mode";
-        case Keyword::PROPS:                return "props";
-        case Keyword::RAILGROUPS:           return "railgroups";
-        case Keyword::RESCUER:              return "rescuer";
-        case Keyword::RIGIDIFIERS:          return "rigidifiers";
-        case Keyword::ROLLON:               return "rollon";
-        case Keyword::ROPABLES:             return "ropables";
-        case Keyword::ROPES:                return "ropes";
-        case Keyword::ROTATORS:             return "rotators";
-        case Keyword::ROTATORS2:            return "rotators2";
-        case Keyword::SCREWPROPS:           return "screwprops";
-        case Keyword::SECTION:              return "section";
-        case Keyword::SECTIONCONFIG:        return "sectionconfig";
-        case Keyword::SET_BEAM_DEFAULTS:    return "set_beam_defaults";
-        case Keyword::SET_BEAM_DEFAULTS_SCALE: return "set_beam_defaults_scale";
-        case Keyword::SET_COLLISION_RANGE:  return "set_collision_range";
-        case Keyword::SET_DEFAULT_MINIMASS: return "set_default_minimass";
-        case Keyword::SET_INERTIA_DEFAULTS: return "set_inertia_defaults";
-        case Keyword::SET_MANAGEDMATERIALS_OPTIONS: return "set_managedmaterials_options";
-        case Keyword::SET_NODE_DEFAULTS:    return "set_node_defaults";
-        case Keyword::SET_SHADOWS:          return "set_shadows";
-        case Keyword::SET_SKELETON_SETTINGS: return "set_skeleton_settings";
-        case Keyword::SHOCKS:               return "shocks";
-        case Keyword::SHOCKS2:              return "shocks2";
-        case Keyword::SHOCKS3:              return "shocks3";
-        case Keyword::SLIDENODE_CONNECT_INSTANTLY: return "slidenode_connect_instantly";
-        case Keyword::SLIDENODES:           return "slidenodes";
-        case Keyword::SLOPE_BRAKE:          return "SlopeBrake";
-        case Keyword::SOUNDSOURCES:         return "soundsources";
-        case Keyword::SOUNDSOURCES2:        return "soundsources2";
-        case Keyword::SPEEDLIMITER:         return "speedlimiter";
-        case Keyword::SUBMESH:              return "submesh";
-        case Keyword::SUBMESH_GROUNDMODEL:  return "submesh_groundmodel";
-        case Keyword::TEXCOORDS:            return "texcoords";
-        case Keyword::TIES:                 return "ties";
-        case Keyword::TORQUECURVE:          return "torquecurve";
-        case Keyword::TRACTIONCONTROL:      return "tractioncontrol";
-        case Keyword::TRANSFERCASE:         return "transfercase";
-        case Keyword::TRIGGERS:             return "triggers";
-        case Keyword::TURBOJETS:            return "turbojets";
-        case Keyword::TURBOPROPS:           return "turboprops";
-        case Keyword::TURBOPROPS2:          return "turboprops2";
-        case Keyword::VIDEOCAMERA:          return "videocamera";
-        case Keyword::WHEELDETACHERS:       return "wheeldetachers";
-        case Keyword::WHEELS:               return "wheels";
-        case Keyword::WHEELS2:              return "wheels2";
-        case Keyword::WINGS:                return "wings";
-
-        default:                           return "";
-    }
-}
-
 Document::Module::Module(Ogre::String const & name):
     name(name)
 {}
