+++ source/main/resources/rig_def_fileformat/RigDef_File.h
@@ -61,134 +61,6 @@ extern const char* ROOT_MODULE_NAME;
 // --------------------------------
 // Enums which only carry value
 
-// IMPORTANT! If you add a value here, you must also modify Regexes::IDENTIFY_KEYWORD, it relies on numeric values of this enum.
-enum class Keyword
-{
-    INVALID = 0,
-
-    ADD_ANIMATION = 1,
-    AIRBRAKES,
-    ANIMATORS,
-    ANTILOCKBRAKES,
-    ASSETPACKS,
-    AUTHOR,
-    AXLES,
-    BACKMESH,
-    BEAMS,
-    BRAKES,
-    CAB,
-    CAMERARAIL,
-    CAMERAS,
-    CINECAM,
-    COLLISIONBOXES,
-    COMMANDS,
-    COMMANDS2,
-    COMMENT,
-    CONTACTERS,
-    CRUISECONTROL,
-    DEFAULT_SKIN,
-    DESCRIPTION,
-    DETACHER_GROUP,
-    DISABLEDEFAULTSOUNDS,
-    ENABLE_ADVANCED_DEFORMATION,
-    END,
-    END_COMMENT,
-    END_DESCRIPTION,
-    END_SECTION,
-    ENGINE,
-    ENGOPTION,
-    ENGTURBO,
-    ENVMAP,
-    EXHAUSTS,
-    EXTCAMERA,
-    FILEFORMATVERSION,
-    FILEINFO,
-    FIXES,
-    FLARES,
-    FLARES2,
-    FLARES3,
-    FLAREGROUPS_NO_IMPORT,
-    FLEXBODIES,
-    FLEXBODY_CAMERA_MODE,
-    FLEXBODYWHEELS,
-    FORSET,
-    FORVERT,
-    FORWARDCOMMANDS,
-    FUSEDRAG,
-    GLOBALS,
-    GUID,
-    GUISETTINGS,
-    HELP,
-    HIDEINCHOOSER,
-    HOOKGROUP, // obsolete, ignored
-    HOOKS,
-    HYDROS,
-    IMPORTCOMMANDS,
-    INTERAXLES,
-    LOCKGROUPS,
-    LOCKGROUP_DEFAULT_NOLOCK,
-    MANAGEDMATERIALS,
-    MATERIALFLAREBINDINGS,
-    MESHWHEELS,
-    MESHWHEELS2,
-    MINIMASS,
-    NODECOLLISION, // obsolete
-    NODES,
-    NODES2,
-    PARTICLES,
-    PISTONPROPS,
-    PROP_CAMERA_MODE,
-    PROPS,
-    RAILGROUPS,
-    RESCUER,
-    RIGIDIFIERS, // obsolete
-    ROLLON,
-    ROPABLES,
-    ROPES,
-    ROTATORS,
-    ROTATORS2,
-    SCREWPROPS,
-    SCRIPTS,
-    SECTION,
-    SECTIONCONFIG,
-    SET_BEAM_DEFAULTS,
-    SET_BEAM_DEFAULTS_SCALE,
-    SET_COLLISION_RANGE,
-    SET_DEFAULT_MINIMASS,
-    SET_INERTIA_DEFAULTS,
-    SET_MANAGEDMATERIALS_OPTIONS,
-    SET_NODE_DEFAULTS,
-    SET_SHADOWS,
-    SET_SKELETON_SETTINGS,
-    SHOCKS,
-    SHOCKS2,
-    SHOCKS3,
-    SLIDENODE_CONNECT_INSTANTLY,
-    SLIDENODES,
-    SLOPE_BRAKE,
-    SOUNDSOURCES,
-    SOUNDSOURCES2,
-    SPEEDLIMITER,
-    SUBMESH,
-    SUBMESH_GROUNDMODEL,
-    TEXCOORDS,
-    TIES,
-    TORQUECURVE,
-    TRACTIONCONTROL,
-    TRANSFERCASE,
-    TRIGGERS,
-    TURBOJETS,
-    TURBOPROPS,
-    TURBOPROPS2,
-    VIDEOCAMERA,
-    WHEELDETACHERS,
-    WHEELS,
-    WHEELS2,
-    WINGS
-};
-
-const char* KeywordToString(Keyword keyword);
-
 enum class SpecialProp
 {
     NONE,
